Returning the right default for south.
We have a tricky situation here. If we want the default for the class,
it should be of the form '0.0 GBP', however, if we want the default
for south (which calls the field.get_default() method), we need to
store only the numerical value 0.0

So this offer this ability, using introspect to understand if the call
comes from south and act appropriately.