changes: MultiValueField instead of DecimalField and fix #8
If you want to use localisation for the form field, you'll run into an `ValueError`: too many values to unpack
This happens due to the fact, that django's `DecimalField` will use a `TextInput` rather than a `NumberInput` if `Field.localize` is set `True`.
In this case, the default value must be a `String` rather than a `Float`.